chore: change app navigation to more flexible: renamed main package into home
