+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/common/FolderHeader.kt
@@ -1,0 +1,24 @@
+package com.wire.android.ui.home.conversations.common
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.wire.android.ui.theme.title03
+import com.wire.android.ui.theme.wireColorScheme
+
+
+@Composable
+fun FolderHeader(name: String) {
+    Text(
+        text = name,
+        modifier = Modifier.padding(start = 16.dp, top = 8.dp, bottom = 8.dp),
+        maxLines = 1,
+        overflow = TextOverflow.Ellipsis,
+        style = MaterialTheme.typography.title03,
+        color = MaterialTheme.wireColorScheme.labelText
+    )
+}
