+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/model/Call.kt
@@ -1,0 +1,19 @@
+package com.wire.android.ui.home.conversations.call.model
+
+import androidx.annotation.DrawableRes
+import com.wire.android.R
+import com.wire.android.ui.home.conversations.all.model.Conversation
+
+data class Call(val callInfo: CallInfo, val conversation: Conversation)
+
+data class CallTime(val date: String, val time: String) {
+    fun toLabel(): String {
+        return "$date, $time"
+    }
+}
+
+data class CallInfo(val callTime: CallTime, val callEvent: CallEvent)
+
+enum class CallEvent(@DrawableRes val drawableResourceId: Int) {
+    MissedCall(R.drawable.ic_missed_call), OutgoingCall(R.drawable.ic_outgoing_call), NoAnswerCall(R.drawable.ic_no_answer_call)
+}
