+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/model/NewActivity.kt
@@ -1,0 +1,13 @@
+package com.wire.android.ui.home.conversations.all.model
+
+data class NewActivity(
+    val eventType: EventType,
+    val conversation: Conversation,
+)
+
+sealed class EventType {
+    data class UnreadMessage(val unreadMessageCount: Int) : EventType()
+    object UnreadMention : EventType()
+    object UnreadReply : EventType()
+    object MissedCall : EventType()
+}
