+++ app/src/main/kotlin/com/wire/android/ui/home/settings/SettingsScreen.kt
@@ -1,0 +1,16 @@
+package com.wire.android.ui.home.settings
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.wire.android.ui.common.UnderConstructionScreen
+
+@Composable
+fun SettingsScreen() {
+    UnderConstructionScreen(screenName = "SettingsScreen")
+}
+
+@Preview(showBackground = false)
+@Composable
+fun SettingsScreenPreview() {
+    SettingsScreen()
+}
