+++ app/src/main/kotlin/com/wire/android/ui/home/support/SupportScreen.kt
@@ -1,0 +1,16 @@
+package com.wire.android.ui.home.support
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.wire.android.ui.common.UnderConstructionScreen
+
+@Composable
+fun SupportScreen() {
+    UnderConstructionScreen(screenName = "SupportScreen")
+}
+
+@Preview(showBackground = false)
+@Composable
+fun SupportScreenPreview() {
+    SupportScreen()
+}
