+++ app/src/main/kotlin/com/wire/android/ui/home/userprofile/UserProfileScreen.kt
@@ -1,0 +1,16 @@
+package com.wire.android.ui.home.userprofile
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.wire.android.ui.common.UnderConstructionScreen
+
+@Composable
+fun UserProfileScreen() {
+    UnderConstructionScreen(screenName = "UserProfileScreen")
+}
+
+@Preview(showBackground = false)
+@Composable
+fun UserProfileScreenPreview() {
+    UserProfileScreen()
+}
