+++ app/src/main/kotlin/com/wire/android/ui/home/vault/VaultScreen.kt
@@ -1,0 +1,16 @@
+package com.wire.android.ui.home.vault
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.wire.android.ui.common.UnderConstructionScreen
+
+@Composable
+fun VaultScreen() {
+    UnderConstructionScreen(screenName = "VaultScreen")
+}
+
+@Preview(showBackground = false)
+@Composable
+fun VaultScreenPreview() {
+    VaultScreen()
+}
