--- app/src/main/kotlin/com/wire/android/ui/main/archive/ArchiveScreen.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.archive
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun ArchiveScreen() {
-    UnderConstructionScreen(screenName = "ArchiveScreen")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun ArchiveScreenPreview() {
-    ArchiveScreen()
-}
