--- app/src/main/kotlin/com/wire/android/ui/main/conversations/common/Extension.kt
@@ -1,19 +1,0 @@
-package com.wire.android.ui.main.conversations.common
-
-import androidx.compose.foundation.lazy.LazyListScope
-import androidx.compose.foundation.lazy.items
-import androidx.compose.runtime.Composable
-
-
-inline fun <T> LazyListScope.folderWithElements(
-    crossinline header: @Composable () -> String,
-    items: List<T>,
-    crossinline factory: @Composable (T) -> Unit
-) {
-    if (items.isNotEmpty()) {
-        item { FolderHeader(name = header()) }
-        items(items) {
-            factory(it)
-        }
-    }
-}
