--- app/src/main/kotlin/com/wire/android/ui/main/conversations/common/MembershipQualifier.kt
@@ -1,33 +1,0 @@
-package com.wire.android.ui.main.conversations.common
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.wire.android.ui.theme.label03
-
-
-@Composable
-fun MembershipQualifier(label: String) {
-    Text(
-        text = label,
-        color = MaterialTheme.colorScheme.primary,
-        style = MaterialTheme.typography.label03.copy(textAlign = TextAlign.Center),
-        modifier = Modifier
-            .wrapContentWidth()
-            .background(
-                color = MaterialTheme.colorScheme.primaryContainer,
-                shape = RoundedCornerShape(8.dp)
-            )
-            .padding(
-                horizontal = 4.dp,
-                vertical = 2.dp
-            )
-    )
-}
