--- app/src/main/kotlin/com/wire/android/ui/main/settings/SettingsScreen.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.settings
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun SettingsScreen() {
-    UnderConstructionScreen(screenName = "SettingsScreen")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun SettingsScreenPreview() {
-    SettingsScreen()
-}
