--- app/src/main/kotlin/com/wire/android/ui/main/support/SupportScreen.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.support
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun SupportScreen() {
-    UnderConstructionScreen(screenName = "SupportScreen")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun SupportScreenPreview() {
-    SupportScreen()
-}
