--- app/src/main/kotlin/com/wire/android/ui/main/vault/VaultScreen.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.vault
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun VaultScreen() {
-    UnderConstructionScreen(screenName = "VaultScreen")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun VaultScreenPreview() {
-    VaultScreen()
-}
